refactor: adapt to the latest pom structure
