+++ packages/insomnia-smoke-test/playwright/pages/index.ts
@@ -1,1 +1,9 @@
 export { InsomniaApp } from './insomnia-app';
+export { ProjectPage } from './project';
+export { WorkspacePage } from './workspace';
+export { PreferencesPage } from './preferences';
+export { PreferencesDataTab } from './preferences/data-tab';
+
+// Components
+export { StatusbarComponent } from './components/statusbar';
+export { WorkspaceListComponent } from './project/workspace-list';
