+++ packages/insomnia-smoke-test/playwright/pages/insomnia-app.ts
@@ -1,7 +1,9 @@
 import type { ElectronApplication, Page } from '@playwright/test';
 
 import { StatusbarComponent } from './components/statusbar';
+import { PreferencesPage } from './preferences';
 import { ProjectPage } from './project';
+import { WorkspacePage } from './workspace';
 
 /**
  * Root facade for the Insomnia E2E Page Object Model.
@@ -13,6 +15,9 @@ import { ProjectPage } from './project';
  *
  *  // Shared components (statusbar is always present)
  *  await insomnia.statusbar.openPreferences();
+ *
+ *  // Preferences and export
+ *  await insomnia.preferencesPage.dataTab.clickExportProjectButton('My Project');
  * });
  * ```
  *
@@ -20,10 +25,13 @@ import { ProjectPage } from './project';
  *
  * ```
  * InsomniaApp (root)
- * ├── .statusbar     -> StatusbarComponent (convenience shortcut)
- * └── .projectPage     -> ProjectPage
- *     ├── .sidebar      -> ProjectSidebarComponent
- *     └── .workspaceList -> WorkspaceListComponent
+ * ├── .statusbar         -> StatusbarComponent (convenience shortcut)
+ * ├── .projectPage       -> ProjectPage
+ * │   ├── .sidebar      -> ProjectSidebarComponent
+ * │   └── .workspaceList -> WorkspaceListComponent
+ * ├── .workspacePage     -> WorkspacePage
+ * └── .preferencesPage   -> PreferencesPage
+ *     └── .dataTab       -> PreferencesDataTab
  * ```
  */
 export class InsomniaApp {
@@ -41,6 +49,12 @@ export class InsomniaApp {
   /** Project page (project/file list). */
   readonly projectPage: ProjectPage;
 
+  /** Workspace page (debug view). */
+  readonly workspacePage: WorkspacePage;
+
+  /** Preferences page (settings modal). */
+  readonly preferencesPage: PreferencesPage;
+
   constructor(
     readonly page: Page,
     readonly app: ElectronApplication,
@@ -50,6 +64,8 @@ export class InsomniaApp {
 
     // Pages
     this.projectPage = new ProjectPage(page, app);
+    this.workspacePage = new WorkspacePage(page, app);
+    this.preferencesPage = new PreferencesPage(page, app);
   }
 
   // ===========================================================================
